html: rewrite render as visitor
Previously, rendering was implemented as manual walk over the sanity
node tree. Now, there is a Visitor interface that manages the walk and
render uses the visitor interface.

The goal of the change is to make it possible for library clients to
walk the tree. For example, it would be nice to walk a form and check
that the name of all inputs match what the server expects when in the
form action handler.

Using dynamic dispatch when rendering slows down the render slightly.
But, while adding the visitor I noticed that using any instead of []Node
was causing an extra allocation for each tag. Swapping to []Node removed
the extra allocation and more than made up for the cost of the dynamic
dispatch.