html: split the tag and visitor interfaces
Because the visitor requires the caller to walk the tree, it is possible
to split the visitor interfaces. It is still desirable for a single
struct to implement both interfaces because it allows walking the entire
tree with a single allocation, instead of needing to allocate for each
level of the walk.